+++ thinking-in-spring/dependency-injection/src/main/java/org/geekbang/thinking/in/spring/ioc/dependency/injection/constructor/AnnotationDependencyConstructorInjectionDemo.java
@@ -14,32 +14,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.geekbang.thinking.in.spring.ioc.dependency.injection;
+package org.geekbang.thinking.in.spring.ioc.dependency.injection.constructor;
 
+import org.geekbang.thinking.in.spring.ioc.dependency.injection.UserHolder;
 import org.geekbang.thinking.in.spring.ioc.overview.domain.User;
-import org.springframework.beans.factory.config.BeanDefinition;
-import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.xml.XmlBeanDefinitionReader;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Bean;
 
 /**
- * 基于 API 实现依赖 Setter 方法注入示例
+ * 基于 Java 注解的依赖 Constructor 注入示例
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  * @since
  */
-public class ApiDependencySetterInjectionDemo {
+public class AnnotationDependencyConstructorInjectionDemo {
 
     public static void main(String[] args) {
 
         // 创建 BeanFactory 容器
         AnnotationConfigApplicationContext applicationContext = new AnnotationConfigApplicationContext();
-
-        // 生成 UserHolder 的 BeanDefinition
-        BeanDefinition userHolderBeanDefinition = createUserHolderBeanDefinition();
-        // 注册 UserHolder 的 BeanDefinition
-        applicationContext.registerBeanDefinition("userHolder", userHolderBeanDefinition);
+        // 注册 Configuration Class（配置类）
+        applicationContext.register(AnnotationDependencyConstructorInjectionDemo.class);
 
         XmlBeanDefinitionReader beanDefinitionReader = new XmlBeanDefinitionReader(applicationContext);
 
@@ -58,21 +54,8 @@ public class ApiDependencySetterInjectionDemo {
         applicationContext.close();
     }
 
-    /**
-     * 为 {@link UserHolder} 生成 {@link BeanDefinition}
-     *
-     * @return
-     */
-    private static BeanDefinition createUserHolderBeanDefinition() {
-        BeanDefinitionBuilder definitionBuilder = BeanDefinitionBuilder.genericBeanDefinition(UserHolder.class);
-        definitionBuilder.addPropertyReference("user", "superUser");
-        return definitionBuilder.getBeanDefinition();
+    @Bean
+    public UserHolder userHolder(User user) {
+        return new UserHolder(user);
     }
-
-//    @Bean
-//    public UserHolder userHolder(User user) { // superUser -> primary = true
-//        UserHolder userHolder = new UserHolder();
-//        userHolder.setUser(user);
-//        return userHolder;
-//    }
 }
