+++ thinking-in-spring/dependency-injection/src/main/java/org/geekbang/thinking/in/spring/ioc/dependency/injection/field/AnnotationDependencyFieldInjectionDemo.java
@@ -14,27 +14,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.geekbang.thinking.in.spring.ioc.dependency.injection;
+package org.geekbang.thinking.in.spring.ioc.dependency.injection.field;
 
+import org.geekbang.thinking.in.spring.ioc.dependency.injection.UserHolder;
 import org.geekbang.thinking.in.spring.ioc.overview.domain.User;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.xml.XmlBeanDefinitionReader;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Bean;
 
+import javax.annotation.Resource;
+
 /**
- * 基于 Java 注解的依赖 Constructor 注入示例
+ * 基于 Java 注解的依赖字段注入示例
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  * @since
  */
-public class AnnotationDependencyConstructorInjectionDemo {
+public class AnnotationDependencyFieldInjectionDemo {
+
+    @Autowired
+    private
+//    static // @Autowired 会忽略掉静态字段
+            UserHolder userHolder;
+
+    @Resource
+    private UserHolder userHolder2;
 
     public static void main(String[] args) {
 
         // 创建 BeanFactory 容器
         AnnotationConfigApplicationContext applicationContext = new AnnotationConfigApplicationContext();
-        // 注册 Configuration Class（配置类）
-        applicationContext.register(AnnotationDependencyConstructorInjectionDemo.class);
+        // 注册 Configuration Class（配置类） -> Spring Bean
+        applicationContext.register(AnnotationDependencyFieldInjectionDemo.class);
 
         XmlBeanDefinitionReader beanDefinitionReader = new XmlBeanDefinitionReader(applicationContext);
 
@@ -45,9 +57,16 @@ public class AnnotationDependencyConstructorInjectionDemo {
         // 启动 Spring 应用上下文
         applicationContext.refresh();
 
-        // 依赖查找并且创建 Bean
-        UserHolder userHolder = applicationContext.getBean(UserHolder.class);
+        // 依赖查找 AnnotationDependencyFieldInjectionDemo Bean
+        AnnotationDependencyFieldInjectionDemo demo = applicationContext.getBean(AnnotationDependencyFieldInjectionDemo.class);
+
+        // @Autowired 字段关联
+        UserHolder userHolder = demo.userHolder;
         System.out.println(userHolder);
+        System.out.println(demo.userHolder2);
+
+        System.out.println(userHolder == demo.userHolder2);
+
 
         // 显示地关闭 Spring 应用上下文
         applicationContext.close();
