+++ thinking-in-spring/dependency-injection/src/main/java/org/geekbang/thinking/in/spring/ioc/dependency/injection/setter/ApiDependencySetterInjectionDemo.java
@@ -14,49 +14,33 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.geekbang.thinking.in.spring.ioc.dependency.injection;
+package org.geekbang.thinking.in.spring.ioc.dependency.injection.setter;
 
+import org.geekbang.thinking.in.spring.ioc.dependency.injection.UserHolder;
 import org.geekbang.thinking.in.spring.ioc.overview.domain.User;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.xml.XmlBeanDefinitionReader;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Bean;
 
-import javax.annotation.Resource;
-
 /**
- * 基于 Java 注解的依赖方法注入示例
+ * 基于 API 实现依赖 Setter 方法注入示例
  *
  * @author <a href="mailto:mercyblitz@gmail.com">Mercy</a>
  * @since
  */
-public class AnnotationDependencyMethodInjectionDemo {
-
-    private UserHolder userHolder;
-
-    private UserHolder userHolder2;
-
-    @Autowired
-    public void init1(UserHolder userHolder) {
-        this.userHolder = userHolder;
-    }
-
-    @Resource
-    public void init2(UserHolder userHolder2) {
-        this.userHolder2 = userHolder2;
-    }
-
-    @Bean
-    public UserHolder userHolder(User user) {
-        return new UserHolder(user);
-    }
+public class ApiDependencySetterInjectionDemo {
 
     public static void main(String[] args) {
 
         // 创建 BeanFactory 容器
         AnnotationConfigApplicationContext applicationContext = new AnnotationConfigApplicationContext();
-        // 注册 Configuration Class（配置类） -> Spring Bean
-        applicationContext.register(AnnotationDependencyMethodInjectionDemo.class);
+
+        // 生成 UserHolder 的 BeanDefinition
+        BeanDefinition userHolderBeanDefinition = createUserHolderBeanDefinition();
+        // 注册 UserHolder 的 BeanDefinition
+        applicationContext.registerBeanDefinition("userHolder", userHolderBeanDefinition);
 
         XmlBeanDefinitionReader beanDefinitionReader = new XmlBeanDefinitionReader(applicationContext);
 
@@ -67,19 +51,29 @@ public class AnnotationDependencyMethodInjectionDemo {
         // 启动 Spring 应用上下文
         applicationContext.refresh();
 
-        // 依赖查找 AnnotationDependencyFieldInjectionDemo Bean
-        AnnotationDependencyMethodInjectionDemo demo = applicationContext.getBean(AnnotationDependencyMethodInjectionDemo.class);
-
-        // @Autowired 字段关联
-        UserHolder userHolder = demo.userHolder;
+        // 依赖查找并且创建 Bean
+        UserHolder userHolder = applicationContext.getBean(UserHolder.class);
         System.out.println(userHolder);
-        System.out.println(demo.userHolder2);
-
-        System.out.println(userHolder == demo.userHolder2);
-
 
         // 显示地关闭 Spring 应用上下文
         applicationContext.close();
     }
 
+    /**
+     * 为 {@link UserHolder} 生成 {@link BeanDefinition}
+     *
+     * @return
+     */
+    private static BeanDefinition createUserHolderBeanDefinition() {
+        BeanDefinitionBuilder definitionBuilder = BeanDefinitionBuilder.genericBeanDefinition(UserHolder.class);
+        definitionBuilder.addPropertyReference("user", "superUser");
+        return definitionBuilder.getBeanDefinition();
+    }
+
+//    @Bean
+//    public UserHolder userHolder(User user) { // superUser -> primary = true
+//        UserHolder userHolder = new UserHolder();
+//        userHolder.setUser(user);
+//        return userHolder;
+//    }
 }
