+++ thinking-in-spring/dependency-injection/src/main/java/org/geekbang/thinking/in/spring/ioc/dependency/injection/setter/AutoWiringByNameDependencySetterInjectionDemo.java
@@ -14,8 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.geekbang.thinking.in.spring.ioc.dependency.injection;
+package org.geekbang.thinking.in.spring.ioc.dependency.injection.setter;
 
+import org.geekbang.thinking.in.spring.ioc.dependency.injection.UserHolder;
 import org.springframework.beans.factory.support.DefaultListableBeanFactory;
 import org.springframework.beans.factory.xml.XmlBeanDefinitionReader;
